[layout] Switch radio/checkbox to have an intrinsic inline/block size.
... as opposed to setting an explicit width/height when we have an
effective appearance, and the width/height is auto.

This is effectively what the code in layout_theme_default was doing,
but this is in a less broken way.

This has the primary side effect of allowing radio/checkbox to respect
stretching (in flex & grid).

Most of the image differences come at non-integer zooms. E.g.
Previously we'd round down(?) for things like:

13px * 1.5 = 19px
now:
13px * 1.5 = 19.5px

Bug: 768999
Change-Id: I222f07ada71ea56a3656af03b459d7a5007e924d